Live Data
Added an option to the Cutoff, Cone, Planet, Trajectory, and Trace functions to use live data hosted on online servers (NOAA, WDC Kyoto, and GFZ-Potsdam) instead of user inputs. It is only available for 7 days before the current date of OTSO's use. This could be expanded in future. This functionality allows for semi-realtime computations with OTSO.